Added loggers.

+++ TextShareApi/Controllers/TextController.cs
@@ -1,4 +1,3 @@
-using System.Diagnostics;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using TextShareApi.ClassesLib;
@@ -6,7 +5,6 @@ using TextShareApi.Dtos.Text;
 using TextShareApi.Extensions;
 using TextShareApi.Interfaces.Services;
 using TextShareApi.Mappers;
-using TextShareApi.Models;
 using TextShareApi.Models.Enums;
 
 namespace TextShareApi.Controllers;
@@ -14,30 +12,35 @@ namespace TextShareApi.Controllers;
 [Route("api/[controller]")]
 [ApiController]
 public class TextController : ControllerBase {
-    private readonly ITextService _textService; 
+    private readonly ILogger<TextController> _logger;
+    private readonly ITextService _textService;
 
-    public TextController(ITextService textService) {
+    public TextController(ITextService textService,
+        ILogger<TextController> logger) {
         _textService = textService;
+        _logger = logger;
     }
 
     [HttpPost]
     [Authorize]
     public async Task<IActionResult> Create() {
-        if (!ModelState.IsValid) {
-            return BadRequest(ModelState);
+        using var sectionTimer = SectionTimer.StartNew(_logger);
+
+        if (!ModelState.IsValid) return BadRequest(ModelState);
+
+        var senderName = User.GetUserName();
+        if (senderName == null) {
+            _logger.LogCritical("Sender name is null");
+            throw new ArgumentNullException(nameof(senderName));
         }
 
-        string? senderName = User.GetUserName(); // TODO: Fix to return nullable type.
-        Debug.Assert(senderName != null);
-
-        Result<Text> creationResult = await _textService.Create(senderName);
+        var creationResult = await _textService.Create(senderName);
         if (!creationResult.IsSuccess) {
-            if (creationResult.IsClientError) {
-                return BadRequest(creationResult.Error);
-            }
+            if (creationResult.IsClientError) return BadRequest(creationResult.Error);
             return StatusCode(500, creationResult.Error);
         }
 
+        sectionTimer.SetMessage($"Created text: {creationResult.Value}");
         return CreatedAtAction(nameof(GetById),
             new { id = creationResult.Value.Id },
             creationResult.Value.ToTextDto());
@@ -45,52 +48,57 @@ public class TextController : ControllerBase {
 
     [HttpGet("{id}")]
     public async Task<IActionResult> GetById([FromRoute] string id, [FromQuery] string? requestPassword) {
-        string? senderName = User.GetUserName();
+        using var sectionTimer = SectionTimer.StartNew(_logger);
 
-        for (int i = 0; i != 10; i++) {
-            Debug.WriteLine("From Debug!");
-        }
-        
-        Result<Text> getResult = await _textService.GetById(id, senderName, requestPassword);
+        var senderName = User.GetUserName();
+
+        var getResult = await _textService.GetById(id, senderName, requestPassword);
         if (!getResult.IsSuccess) {
-            if (getResult.IsClientError) {
-                return Forbid();
-            }
+            if (getResult.IsClientError) return Forbid();
             return StatusCode(500, getResult.Error);
         }
-        
+
+        sectionTimer.SetMessage($"Retrieved text: {getResult.Value.Id}");
+
         return Ok(getResult.Value.ToTextDto());
     }
 
     [HttpGet("/myTexts")]
     [Authorize]
     public async Task<IActionResult> GetMyTexts() {
-        string? senderName = User.GetUserName();
-        Debug.Assert(senderName != null);
-        
-        Result<List<Text>> getResult = await _textService.GetAccountTexts(senderName);
+        using var sectionTimer = SectionTimer.StartNew(_logger);
+
+        var senderName = User.GetUserName();
+        if (senderName == null) {
+            _logger.LogCritical("Sender name is null");
+            throw new ArgumentNullException(nameof(senderName));
+        }
+
+        var getResult = await _textService.GetAccountTexts(senderName);
         if (!getResult.IsSuccess) {
-            if (getResult.IsClientError) {
-                return Forbid();
-            }
+            if (getResult.IsClientError) return Forbid();
             return StatusCode(500, getResult.Error);
         }
-        
+
+        sectionTimer.SetMessage($"Returned texts: {getResult.Value.Count}");
+
         return Ok(getResult.Value.Select(t => t.ToTextWithoutContentDto()).ToList());
     }
-    
+
     [HttpGet]
     public async Task<IActionResult> GetAllAvailable() {
-        string? senderName = User.GetUserName();
-        
-        Result<List<Text>> getResult = await _textService.GetAllAvailable(senderName);
+        using var sectionTimer = SectionTimer.StartNew(_logger);
+
+        var senderName = User.GetUserName();
+
+        var getResult = await _textService.GetAllAvailable(senderName);
         if (!getResult.IsSuccess) {
-            if (getResult.IsClientError) {
-                return BadRequest(getResult.Error);
-            }
+            if (getResult.IsClientError) return BadRequest(getResult.Error);
             return StatusCode(500, getResult.Error);
         }
-        
+
+        sectionTimer.SetMessage($"Returned Texts: {getResult.Value.Count}");
+
         return Ok(getResult.Value.Select(t => t.ToTextWithoutContentDto()).ToList());
     }
 
@@ -99,26 +107,29 @@ public class TextController : ControllerBase {
     public async Task<IActionResult> UpdateContent([FromRoute] string id,
         [FromQuery] string? requestPassword,
         [FromBody] TextContentUpdateDto updateDto) {
-        if (!ModelState.IsValid) {
-            return BadRequest(ModelState.ValidationState);
+        using var sectionTimer = SectionTimer.StartNew(_logger);
+
+        if (!ModelState.IsValid) return BadRequest(ModelState.ValidationState);
+
+        var senderName = User.GetUserName();
+        if (senderName == null) {
+            _logger.LogCritical("Sender name is null");
+            throw new ArgumentNullException(nameof(senderName));
         }
-        
-        string? senderName = User.GetUserName();
-        Debug.Assert(senderName != null);
+
 
         var contentDto = new UpdateTextDto {
             Text = updateDto.Text,
-            UpdatePassword = false,
+            UpdatePassword = false
         };
 
         var updateResult = await _textService.Update(id, senderName, requestPassword, contentDto);
         if (!updateResult.IsSuccess) {
-            if (updateResult.IsClientError) {
-                return Forbid();
-            }
+            if (updateResult.IsClientError) return Forbid();
             return StatusCode(500, updateResult.Error);
         }
-        
+
+        sectionTimer.SetMessage($"Updated text content: {updateResult.Value.Id}");
         return Ok(updateResult.Value.ToTextDto());
     }
 
@@ -127,34 +138,36 @@ public class TextController : ControllerBase {
     public async Task<IActionResult> UpdateSecuritySettings([FromRoute] string id,
         [FromQuery] string? requestPassword,
         [FromBody] TextSecSetUpdateDto secSetsDto) {
-        if (!ModelState.IsValid) {
-            return BadRequest(ModelState.ValidationState);
+        using var sectionTimer = SectionTimer.StartNew(_logger);
+
+        if (!ModelState.IsValid) return BadRequest(ModelState.ValidationState);
+
+        var senderName = User.GetUserName();
+        if (senderName == null) {
+            _logger.LogCritical("Sender name is null");
+            throw new ArgumentNullException(nameof(senderName));
         }
 
-        string? senderName = User.GetUserName();
-        Debug.Assert(senderName != null);
-        
+
         var updateDto = new UpdateTextDto {
             Password = secSetsDto.Password,
-            UpdatePassword = secSetsDto.UpdatePassword,
+            UpdatePassword = secSetsDto.UpdatePassword
         };
 
         if (secSetsDto.AccessType != null) {
-            if (Enum.TryParse(secSetsDto.AccessType, out AccessType newType)) {
+            if (Enum.TryParse(secSetsDto.AccessType, out AccessType newType))
                 updateDto.AccessType = newType;
-            }
-            else {
+            else
                 return BadRequest("Invalid access type");
-            }
         }
-        
-        Result<Text> updateResult = await _textService.Update(id, senderName, requestPassword, updateDto);
+
+        var updateResult = await _textService.Update(id, senderName, requestPassword, updateDto);
         if (!updateResult.IsSuccess) {
-            if (updateResult.IsClientError) {
-                return Forbid();
-            }
+            if (updateResult.IsClientError) return Forbid();
             return StatusCode(500, updateResult.Error);
         }
+
+        sectionTimer.SetMessage($"Updated text security settings: {updateResult.Value.Id}");
 
         return Ok(updateResult.Value.ToTextDto());
     }
@@ -162,16 +175,21 @@ public class TextController : ControllerBase {
     [HttpDelete("{id}")]
     [Authorize]
     public async Task<IActionResult> Delete([FromRoute] string id, [FromQuery] string? requestPassword) {
-        string? senderName = User.GetUserName();
-        Debug.Assert(senderName != null);
+        using var sectionTimer = SectionTimer.StartNew(_logger);
 
-        Result deleteResult = await _textService.Delete(id, senderName, requestPassword);
+        var senderName = User.GetUserName();
+        if (senderName == null) {
+            _logger.LogCritical("Sender name is null");
+            throw new ArgumentNullException(nameof(senderName));
+        }
+
+        var deleteResult = await _textService.Delete(id, senderName, requestPassword);
         if (!deleteResult.IsSuccess) {
-            if (deleteResult.IsClientError) {
-                return Forbid();
-            }
+            if (deleteResult.IsClientError) return Forbid();
             return StatusCode(500, deleteResult.Error);
         }
+
+        sectionTimer.SetMessage($"Deleted Text: {id}");
 
         return NoContent();
     }
